Expand practice questions and cases; refactor exam skills
