+++ index.html
@@ -1286,6 +1286,39 @@
                                             <p class="mt-3 text-xs text-slate-400">Use analytical and signposting words to improve structure (e.g., "Therefore", "Significantly", "However", "For example") [cite: 709-711].</p>
                                         </div>
                                     </div>
+                                    <div class="practice-question">
+                                        <p class="font-medium practice-question-text">6. Explain the significance of Section 109 of the Australian Constitution. (3 marks)</p>
+                                        <button class="text-xs text-purple-600 hover:underline practice-question-toggle-answer">Show/Hide Answer Points</button>
+                                        <div class="practice-question-answer p-2 bg-purple-50 rounded mt-1 hidden">
+                                            <ul class="list-disc ml-5 text-xs">
+                                                <li>Provides that valid Commonwealth law prevails over inconsistent state law.</li>
+                                                <li>Ensures national consistency in concurrent areas.</li>
+                                                <li>Limits state legislative power when conflicts arise.</li>
+                                            </ul>
+                                        </div>
+                                    </div>
+                                    <div class="practice-question">
+                                        <p class="font-medium practice-question-text">7. Describe how Sections 7 and 24 of the Australian Constitution protect the principle of representative government. (3 marks)</p>
+                                        <button class="text-xs text-purple-600 hover:underline practice-question-toggle-answer">Show/Hide Answer Points</button>
+                                        <div class="practice-question-answer p-2 bg-purple-50 rounded mt-1 hidden">
+                                            <ul class="list-disc ml-5 text-xs">
+                                                <li>Require that members of Parliament are directly chosen by the people.</li>
+                                                <li>Imply a right to vote and be represented.</li>
+                                                <li>Allow challenges to laws that unduly restrict this democratic principle.</li>
+                                            </ul>
+                                        </div>
+                                    </div>
+                                    <div class="practice-question">
+                                        <p class="font-medium practice-question-text">8. Explain how the separation of powers acts as a check on the law-making of parliament. (3 marks)</p>
+                                        <button class="text-xs text-purple-600 hover:underline practice-question-toggle-answer">Show/Hide Answer Points</button>
+                                        <div class="practice-question-answer p-2 bg-purple-50 rounded mt-1 hidden">
+                                            <ul class="list-disc ml-5 text-xs">
+                                                <li>Distributes legislative, executive and judicial powers between different bodies.</li>
+                                                <li>Prevents concentration of power in one branch.</li>
+                                                <li>Allows courts to review parliamentary actions where laws conflict with the Constitution.</li>
+                                            </ul>
+                                        </div>
+                                    </div>
 
                                     <div id="bridging-skills-to-knowledge-content" class="primary-exam-skill-content hidden">
                                         <h5 class="text-xl font-semibold text-purple-700 mb-1">Bridging Key Skills to Key Knowledge</h5>
