+++ script.js
@@ -99,18 +99,15 @@ document.addEventListener('DOMContentLoaded', function () {
         if (targetId === 'u4aos1-key-skills-hub' && typeof window.initializeKeySkillsHub === 'function') {
             window.initializeKeySkillsHub();
         }
-        // Call initializeKeySkillsHub when the main "Exam Skills Helper" accordion is opened,
+        // Call initializeExamSkillsHelper when the main "Exam Skills Helper" accordion is opened,
         // as it contains multiple tools that need initialization (Bridging Skills, Templates, Pitfalls).
-        if (targetId === 'u4aos1-exam-skills' && typeof window.initializeKeySkillsHub === 'function') {
-            console.log("DEBUG: script.js - initializeToolIfNeeded: u4aos1-exam-skills detected, attempting to call window.initializeKeySkillsHub.");
-            window.initializeKeySkillsHub();
+        if (targetId === 'u4aos1-exam-skills' && typeof window.initializeExamSkillsHelper === 'function') {
+            console.log("DEBUG: script.js - initializeToolIfNeeded: u4aos1-exam-skills detected, attempting to call window.initializeExamSkillsHelper.");
+            window.initializeExamSkillsHelper();
         }
         if (targetId === 'u4aos1-glossary' && typeof window.setupCategorizedGlossary === 'function') {
             window.setupCategorizedGlossary();
         }
-    }
-
-    function handleU4AOS1ContentToggle(buttonToActivate) {
         const targetId = buttonToActivate.dataset.target;
         
         u4aos1ContentToggles.forEach(btn => {
